Added a new function to create a new part in PySldWrap
This commit adds a new function to the PySldWrap module for creating a new part in SolidWorks. The function takes a path as input and creates a new part at the specified location. This function can be used in conjunction with the existing functions in the module to automate the creation of sketches in SolidWorks.